requirements-scrcpy.txt
# scrcpy-client installed separately to avoid av version conflict
scrcpy-client>=0.4.7
